backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
supabase
python-dotenv
slowapi
google-generativeai
pydantic
python-multipart
resend
httpx